🚚 rename: DashboardContent -> ../
DashboardContent는 templates/Dashboard에서는 사용하지 않는, 관계 없는 파일이기 때문
