+++ app/src/components/templates/Dashboard/index.ts
@@ -1,2 +1,1 @@
 export * from './Dashboard';
-export * from './DashboardContent';
