+++ app/src/pages/HomePage/contents/TigCountPerCoalition.tsx
@@ -1,5 +1,5 @@
 import { Center, H3Text, Text, VStack } from '@components/common';
-import { DashboardContent } from '@components/templates/Dashboard';
+import { DashboardContent } from '@components/templates/DashboardContent';
 import styled from '@emotion/styled';
 import { numberWithUnitFormatter } from '@utils/formatters';
 import dayjs from 'dayjs';
