+++ app/src/pages/ProfileEvalTab/contents/AverageCommentLength.tsx
@@ -6,7 +6,7 @@ import {
   ApolloNotFound,
 } from '@components/elements/DashboardContentView';
 import { NumberDefault } from '@components/elements/DashboardContentView/Text';
-import { DashboardContent } from '@components/templates/Dashboard';
+import { DashboardContent } from '@components/templates/DashboardContent';
 import { useParams } from 'react-router-dom';
 
 const GET_PERSONAL_AVERAGE_COMMENT_LENGTH = gql(/* GraphQL */ `
