+++ app/src/pages/ProfileGeneralTab/contents/PrefferedTime.tsx
@@ -5,7 +5,7 @@ import {
   ApolloBadRequest,
   ApolloNotFound,
 } from '@components/elements/DashboardContentView';
-import { DashboardContent } from '@components/templates/Dashboard';
+import { DashboardContent } from '@components/templates/DashboardContent';
 import styled from '@emotion/styled';
 import { percentFormatter } from '@utils/formatters/percentFormatter';
 import dayjs from 'dayjs';
